+++ src/main/java/chylex/hee/system/weight/IWeightProvider.java
@@ -1,5 +1,5 @@
 package chylex.hee.system.weight;
 
 public interface IWeightProvider{
-	public short getWeight();
+	public int getWeight();
 }
